+++ app/components/ListCardIcon.jsx
@@ -6,7 +6,7 @@ const ListCardIcon = ({ item, type, onClick, isSelected }) => {
   return (
     <div
       onClick={onClick}
-      className={`flex items-center justify-center rounded-md p-1 w-10 h-10 lg:h-8 lg:w-8 ${
+      className={`relative flex items-center justify-center rounded-md p-1 w-10 h-10 lg:h-8 lg:w-8 ${
         !isSelected && "hover:brightness-80"
       }`}
       style={{
@@ -26,7 +26,6 @@ const ListCardIcon = ({ item, type, onClick, isSelected }) => {
               : "black"
             : getTextColor(item?.color?.hex)
         }
-        classes="relative"
         size={18}
       />
     </div>
