+++ app/components/navigation/MobileListViewBreadcrumbs.jsx
@@ -1,49 +1,31 @@
-import { Popover } from "@mantine/core";
+import { useState } from "react";
+import { Tooltip } from "@mantine/core";
 import { BreadcrumbTrail } from "..";
 import { MapPin } from "lucide-react";
-import { useContext } from "react";
-import { ModalContext } from "@/app/providers";
 
 export default function MobileListViewBreadcrumbs({ data }) {
-  const { activePopoverId, setActivePopoverId } = useContext(ModalContext);
+  const [opened, setOpened] = useState(false);
 
-  const opened = activePopoverId === data?.id + "-location";
-  const setOpened = () =>
-    setActivePopoverId(opened ? null : data?.id + "-location");
   return (
-    <Popover
+    <Tooltip
       position="top"
-      classNames={{ dropdown: "w-max !max-w-[95vw] !py-1.5 !px-2.5" }}
-      shadow="0px 4px 8px #00000044"
-      radius="xl"
-      withArrow
-      arrowSize={12}
-      offset={8}
-      closeOnEscape
-      closeOnClickOutside
+      label={<BreadcrumbTrail data={data} showAll={true} />}
       opened={opened}
-      onChange={setOpened}
+      events={{ hover: true, focus: true, touch: true }}
+      classNames={{
+        tooltip: "!bg-white !px-2 !py-1 dropshadow !rounded-full",
+      }}
     >
-      <Popover.Target>
-        <button
-          onClick={setOpened}
-          className={`relative rounded h-8 px-2 !flex !items-center text-center ${
-            data?.location || data?.container || data?.parentContainer
-              ? "hover:brightness-75"
-              : "opacity-50"
-          }`}
-        >
-          <MapPin
-            size={26}
-            fill="var(--mantine-color-primary-5)"
-            className="[&>circle]:fill-white !px-auto shrink-0"
-          />
-        </button>
-      </Popover.Target>
-
-      <Popover.Dropdown>
-        <BreadcrumbTrail data={data} showAll={true} />
-      </Popover.Dropdown>
-    </Popover>
+      <button
+        onClick={setOpened}
+        className="relative rounded h-full min-h-[32px] px-2 flex !items-center text-center [&>svg]:fill-primary-700 hover:brightness-75"
+      >
+        <MapPin
+          size={26}
+          fill="var(--mantine-color-primary-5)"
+          className="[&>circle]:fill-white !px-auto shrink-0"
+        />
+      </button>
+    </Tooltip>
   );
 }
